Reorganized all components as discussed. Closes #16
