+++ src/components/App/redux/reducer.js
