+++ src/components/Dashboard/redux/actions.js
@@ -1,0 +1,1 @@
+// TODO: Create a Dashboard component
