+++ src/components/Dashboard/redux/reducer.js
@@ -1,0 +1,1 @@
+// TODO: Create a Dashboard component
