+++ src/components/Visualization/redux/reducer.js
@@ -1,0 +1,1 @@
+// TODO: Create a Visualization component
